Add component.js to extend point and point-edit subclasses
