+++ src/component.js
@@ -1,0 +1,50 @@
+class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate BaseComponent, only concrete one.`);
+    }
+    this._element = null;
+    this._state = {};
+  }
+
+  get element() {
+    return this._element;
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  createElement(template) {
+    const newElement = document.createElement(`div`);
+    newElement.innerHTML = template;
+    return newElement.firstChild;
+  }
+
+  render() {
+    this._element = this.createElement(this.template);
+    this.createListeners();
+    return this._element;
+  }
+
+  createListeners() {}
+
+  removeListeners() {}
+
+  unrender() {
+    this.removeListeners();
+    this._element.remove();
+    this._element = null;
+  }
+
+  get date() {
+    return {
+      day: this._day,
+      month: this._month,
+      uniqueDay: this._uniqueDay,
+      time: this._time,
+    };
+  }
+
+}
+export default Component;
